On Controller: Fixed show dictionary and added appropriate error, Fixed load dictionary if clause. On Document: changed termDictionary's value to be a short array. On Parse: added a stop boolean reset method, fixed city check, changed all tf values to short, and fixed the corpus dictionary's entry and city dictionary's entries accordingly. On ReadFile: added a corpus sweep (before the initial sweep) to add the cities to the city and corpus dictionaries. On Indexer: Fixed normalized tf support, added posting for cities creation (only once), added a readFirstLines function for the BufferedReaders array, added 2 functions to write and load the dictionary for show (as a sorted string) and added a reset stop function.

+++ src/GeneralClasses/Document.java
@@ -14,7 +14,7 @@ public class Document {
     private String docDate = null;
     private String city = null;
     private ArrayList<String> termList;
-    private HashMap<String,int[]> termDictionary;
+    private HashMap<String,short[]> termDictionary;
     private int max_tf;
     private int uniqueWords;
 
@@ -87,7 +87,7 @@ public class Document {
      * @param termDictionary - the document's term dictionary
      * @param max_tf - ther document's max_tf
      */
-    public void setTfAndTermDictionary(HashMap<String,int[]> termDictionary, int max_tf) {
+    public void setTfAndTermDictionary(HashMap<String,short[]> termDictionary, int max_tf) {
         this.termDictionary = termDictionary;
         uniqueWords = termDictionary.keySet().size();
         this.max_tf = max_tf;
@@ -145,7 +145,7 @@ public class Document {
      * gets the document's term dictionary
      * @return - the document's term dictionary
      */
-    public HashMap<String,int[]> getTermDictionary() {
+    public HashMap<String,short[]> getTermDictionary() {
         return termDictionary;
     }
